Moves configuration to own module
The configuration and handling of dotenv files was originaly in main, it's now in `config.py`.

This follows a singleton pattern and instantiates itself once.

Main now imports the instantiated settings object, and can reference that throughout the application.
